[DF-7335] Win 10 warnings fix.
Former-commit-id: 58626320bc73f1525daf5624803c9b3d7719870b

+++ Sources/Internal/FileSystem/YamlNode.cpp
@@ -862,7 +862,7 @@ bool YamlNode::InitArrayFromVariantType(const VariantType &varType)
     case VariantType::TYPE_COLOR:
         {
             const Color& color = varType.AsColor();
-            InternalSetVector(color.color.data(), color.color.size());
+            InternalSetVector(color.color.data(), static_cast<uint32>(color.color.size()));
         }
         break;
     default:
